DMP→ESP planner/encoder/merge correctness fixes
A batch of correctness fixes to the planner + encoder + merge layers
of the DMP-to-ESP pipeline.

Encoder fixes:
- InfoEncoder/ScptEncoder: emit SCHR with the canonical fopdoc 20-byte
  layout (Padding/RefCount/CompiledSize/VariableCount/Type/Flags) —
  the runtime SCRIPT_HEADER PDB struct uses a divergent layout at
  offsets 0, 12 and 16-19, and emitting in runtime-style was causing
  the engine to read uiLastID as VariableCount, surfacing as 'SCRIPTS:
  Variable ID NNNNNNNN not found' on every result script
- QustEncoder: drop DATA override emission — the runtime byte
  co-locates ESM-authored flags (StartGameEnabled, Allow*) with
  engine state bits (Started, Active, Completed); emitting from a
  DMP capture either wiped master flags or carried runtime state
  (Doc Mitchell's intro never starting; Sunny's quest appearing
  pre-started)
- CellEncoder: emit EDID subrecord only when EditorId is non-empty
  (xEdit's wbCELL: EDID is optional, but must be first when present)
- PackEncoder/WeapEncoder/SounEncoder: drop signed casts on fields
  the schema already treats as unsigned

Merge:
- CellMerger: binary policy — persistent-only DMP captures merge into
  master, anything with non-persistent content replaces it
- LandOverrideBuilder (new): coordinate NAVI override emission with
  NAVM cell-children planning

Planner trivial fixes:
- RecordCatalog → static; ArgumentNullException.ThrowIfNull
  modernization across planner + planned-writer layers

+++ src/FalloutXbox360Utils/Core/Formats/Esm/Plugin/Writers/Encoders/AI/PackEncoder.cs
@@ -41,7 +41,7 @@ public sealed class PackEncoder : IRecordEncoder
         ["Month"] = m => (byte)m.Month,
         ["DayOfWeek"] = m => (byte)m.DayOfWeek,
         ["Date"] = m => (byte)m.Date,
-        ["Time"] = m => (sbyte)m.Time,
+        ["Time"] = m => m.Time,
         ["Duration"] = m => m.Duration,
     };
